# fixed nvidia key sign bug