--- a/docker/build_image.sh
+++ b/docker/build_image.sh
@@ -38,7 +38,7 @@ do
         -g|--gpu)
           MACHINE=gpu
           DOCKER_TAG="pytorch/torchserve:latest-gpu"
-          BASE_IMAGE="nvidia/cuda:10.2-cudnn7-runtime-ubuntu18.04"
+          BASE_IMAGE="nvidia/cuda:10.2-cudnn8-runtime-ubuntu18.04"
           CUDA_VERSION="cu102"
           shift
           ;;